requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.82.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pinecone==7.0.1
pinecone-plugin-interface==0.0.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.32.0
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
